Move stdlib extensions into stdlib_utils
